libs/bot_arbitraje_gemini_version.py
import requests
import pandas as pd
import time
import hmac
import hashlib
from urllib.parse import urlencode
import json

# Configuración de las API
api_key_binance = "TU_API_KEY_BINANCE"
api_secret_binance = "TU_API_SECRET_BINANCE"
api_key_bybit = "TU_API_KEY_BYBIT"
api_secret_bybit = "TU_API_SECRET_BYBIT"

# Función para generar la firma para Binance
def generate_signature(params):
    query_string = urlencode(sorted(params.items()))
    signature = hmac.new(api_secret_binance.encode(), query_string.encode(), hashlib.sha256).hexdigest()
    return signature

# Función para obtener los precios de Binance
def get_prices_binance(symbol):
    timestamp = int(time.time() * 1000)
    params = {'symbol': symbol, 'timestamp': timestamp}
    signature = generate_signature(params)
    url = f'https://api.binance.com/api/v3/ticker/price?symbol={symbol}&timestamp={timestamp}&signature={signature}'
    response = requests.get(url)
    data = response.json()
    return data['price']

def get_prices_bybit(symbol):
    # ... código para obtener los precios de Bybit ...
    """Obtiene el precio actual de un símbolo en Bybit.

    Args:
        symbol: El símbolo del par de trading (por ejemplo, "BTCUSDT").

    Returns:
        El precio actual del símbolo.
    """

    # Parámetros de la solicitud
    params = {
        "symbol": symbol
    }

    # Generar la firma
    timestamp = int(time.time() * 1000)
    params["timestamp"] = timestamp
    signature = hmac.new(api_secret_bybit.encode(), str(timestamp).encode(), hashlib.sha256).hexdigest()

    # URL de la solicitud
    url = "https://api.bybit.com/v2/public/tickers"

    # Hacer la solicitud HTTP
    headers = {
        "X-MBX-API-KEY": api_key_bybit,
        "timestamp": str(timestamp),
        "signature": signature
    }
    response = requests.get(url, params=params, headers=headers)
    data = json.loads(response.text)

    # Buscar el precio del símbolo en la respuesta
    for result in data['result']:
        if result['symbol'] == symbol:
            return result['lastPrice']

    raise Exception("No se encontró el símbolo")

# Función para obtener los precios de múltiples pares de criptomonedas
def get_prices(symbols):
    # ... código para obtener los precios de Binance y Bybit para múltiples símbolos ...
    return prices_df

def calcular_liquidez(df_historico):
  """
  Calcula la liquidez de un símbolo basándose en el volumen de negociación de las últimas 24 horas.

  Args:
    df_historico: DataFrame con los datos históricos de precios y volumen.
    simbolo: El símbolo del par de criptomonedas.

  Returns:
    Un valor numérico que representa la liquidez del símbolo.
  """
  volumen_promedio_24h = {}
  for simbolo in df_historico.index:
      volumen_promedio_24h[simbolo] = df_historico.loc[simbolo, 'volumen'].tail(24).mean()
  # Calcular el volumen promedio de las últimas 24 horas
  

  # Normalizar el volumen (opcional)
  # Puedes normalizar el volumen en función de otros factores, como el precio promedio o el capitalización de mercado.

  return volumen_promedio_24h

def calcular_costo_de_transaccion(precios_historicos):
    fee = {}
    for simbolo in precios_historicos.index:
        fee[simbolo] = .02

    return fee

def calcular_umbral(df_historico, liquidez, volatilidad, costos_transaccion):
  """
  Calcula un umbral dinámico para identificar oportunidades de arbitraje.

  Args:
    df_historico: DataFrame con datos históricos de precios y volumen.
    liquidez: Un diccionario con la liquidez de cada par de criptomonedas.
    volatilidad: Un diccionario con la volatilidad de cada par de criptomonedas.
    costos_transaccion: Un diccionario con los costos de transacción de cada exchange.

  Returns:
    Un diccionario con los umbrales calculados para cada par de criptomonedas.
  """

  umbrales = {}
  for symbol in df_historico.index:
    # Calcular la volatilidad histórica del par
    volatilidad_symbol = df_historico.loc[symbol].std()

    # Calcular el costo total de la transacción (compra y venta)
    costo_transaccion = costos_transaccion['binance'] + costos_transaccion['bybit']

    # Calcular el umbral basado en la volatilidad, liquidez y costos de transacción
    umbral = (volatilidad_symbol * 0.5) + (costo_transaccion / df_historico.loc[symbol, 'precio_promedio']) + (1 / liquidez[symbol])

    umbrales[symbol] = umbral

  return umbrales



# Función para identificar oportunidades de arbitraje
def find_arbitrage_opportunities(price_binance, price_bybit):
    """
    Identifica oportunidades de arbitraje en base a un umbral de diferencia de precios.

    Args:
        prices_df: DataFrame con los precios de los pares de criptomonedas en ambos exchanges.
        threshold: Umbral de diferencia de precios para considerar una oportunidad de arbitraje.

    Returns:
        Lista de pares de criptomonedas con oportunidades de arbitraje.
    """

    umbrales = calcular_umbral(prices_df, calcular_liquidez(prices_df), .5, calcular_costo_de_transaccion(prices_df))

    arbitrage_opportunities = []
    for symbol in prices_df.index:
        threshold = umbrales[symbol]
        price_diff = abs(prices_df.loc[symbol, 'price_binance'] - prices_df.loc[symbol, 'price_bybit'])
        if price_diff / ((prices_df.loc[symbol, 'price_binance'] + prices_df.loc[symbol, 'price_bybit']) / 2) > threshold:
            arbitrage_opportunities.append(symbol)
    return arbitrage_opportunities

# Función para ejecutar una orden
def execute_order(exchange, symbol, side, quantity, price):
    # ... código para enviar una orden al exchange especificado ...

# Bucle principal
while True:
    prices_df = get_prices(['BTCUSDT', 'ETHUSDT', 'BNBUSDT'])  # Lista de pares a monitorear

    # Identificar oportunidades de arbitraje
    arbitrage_opportunities = find_arbitrage_opportunities(prices_df)

    if arbitrage_opportunities:
        print("Oportunidades de arbitraje encontradas:", arbitrage_opportunities)
        # ... lógica para ejecutar órdenes ...

    time.sleep(60)