requirements.txt
binance-connector
numpy
pandas
tensorflow
ta
matplotlib 
intel-tensorflow
scikit-learn
binance-connector
binance-futures-connector
torch
torchvision
torchaudio
websocket
ccxt
tornado

firebase-admin
pdf2image
pytesseract
langchain
openai
langchain-openai
langchain-community 
langchainhub
pydantic
chromadb
bs4